LPS-133931 Rename MappingRichInput to MappingInput because it is a real input and not rich text or editor yet
